+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## [Unreleased]
 
+### Features
+- include short sha in changelog bullet lines (47a312c)
+- changelog markdown generation from parsed commits (#5) (8ac136f)
+
 ### Chores
-- scaffold project (by mito)
+- scaffold project (by mito) (2af19fc)